Handle nested root selection
